[tsne] fix size mismatch warnings

+++ src/tsne-cuda/debug_utils.cu
@@ -45,8 +45,8 @@ void tsne::debug::PrintArray(
     const thrust::device_vector<T>& d_matrix,
     const int N,
     const int M) {
-    for (size_t i = 0; i < N; i++) {
-        for (size_t j = 0; j < M; j++) {
+    for (int i = 0; i < N; i++) {
+        for (int j = 0; j < M; j++) {
             std::cout << d_matrix[i + j * N] << " ";
         }
         std::cout << std::endl;
